AIYHD/requirements.txt
flask
flask-cors
opencv-python-headless
mediapipe
google-generativeai
gunicorn
numpy